requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==23.0.0
